+++ src/PropertyRegistry.php
@@ -4,7 +4,7 @@ namespace SIL;
 
 use SMW\DIProperty;
 
-define( 'SIL_PROP_CONTAINER', 'Has interlanguage links' );
+define( 'SIL_PROP_CONTAINER', 'Has interlanguage link' );
 
 define( 'SIL_PROP_ILL_REF', 'Interlanguage reference' );
 define( 'SIL_PROP_ILL_LANG', 'Page content language' );
@@ -40,56 +40,67 @@ class PropertyRegistry {
 			self::SIL_CONTAINER => array(
 				'label' => SIL_PROP_CONTAINER,
 				'type'  => '__sob',
-				'alias' => wfMessage( 'sil-property-alias-container' )->text(),
+				'alias' => array( wfMessage( 'sil-property-alias-container' )->text(), 'hasInterlanguagelink' ),
 				'visibility' => false,
-				'annotable'  => false
+				'annotableByUser'  => false
 			),
 
 			self::SIL_ILL_LANG => array(
 				'label' => SIL_PROP_ILL_LANG,
 				'type'  => '_txt',
-				'alias' => wfMessage( 'sil-property-ill-alias-language' )->text(),
-				'visibility' => true
+				'alias' => array( wfMessage( 'sil-property-ill-alias-language' )->text() ),
+				'visibility' => true,
+				'annotableByUser'  => true
 			),
 
 			self::SIL_ILL_REF => array(
 				'label' => SIL_PROP_ILL_REF,
 				'type'  => '_wpg',
-				'alias' => wfMessage( 'sil-property-ill-alias-reference' )->text(),
-				'visibility' => true
+				'alias' => array( wfMessage( 'sil-property-ill-alias-reference' )->text() ),
+				'visibility' => true,
+				'annotableByUser'  => false
 			),
 
 			self::SIL_IWL_LANG => array(
 				'label' => SIL_PROP_IWL_LANG,
 				'type'  => '_txt',
-				'alias' => wfMessage( 'sil-property-iwl-alias-language' )->text(),
-				'visibility' => true
+				'alias' => array( wfMessage( 'sil-property-iwl-alias-language' )->text() ),
+				'visibility' => true,
+				'annotableByUser'  => false
 			),
 
 			self::SIL_IWL_REF => array(
 				'label' => SIL_PROP_IWL_REF,
 				'type'  => '_wpg',
-				'alias' => wfMessage( 'sil-property-iwl-alias-reference' )->text(),
-				'visibility' => true
+				'alias' => array( wfMessage( 'sil-property-iwl-alias-reference' )->text() ),
+				'visibility' => true,
+				'annotableByUser'  => false
 			)
 		);
 
 		foreach ( $propertyDefinitions as $propertyId => $definition ) {
-
-			DIProperty::registerProperty(
-				$propertyId,
-				$definition['type'],
-				$definition['label'],
-				$definition['visibility']
-			);
-
-			DIProperty::registerPropertyAlias(
-				$propertyId,
-				$definition['alias']
-			);
+			$this->addPropertyDefinitionFor( $propertyId, $definition  );
 		}
 
 		return true;
+	}
+
+	private function addPropertyDefinitionFor( $propertyId, $definition ) {
+
+		DIProperty::registerProperty(
+			$propertyId,
+			$definition['type'],
+			$definition['label'],
+			$definition['visibility'],
+			$definition['annotableByUser']
+		);
+
+		foreach ( $definition['alias'] as $alias ) {
+			DIProperty::registerPropertyAlias(
+				$propertyId,
+				$alias
+			);
+		}
 	}
 
 }