+++ tests/phpunit/Unit/CacheKeyProviderTest.php
@@ -2,11 +2,10 @@
 
 namespace SIL\Tests;
 
-use SIL\CacheKeyGenerator;
+use SIL\CacheKeyProvider;
 
 /**
- * @covers \SIL\CacheKeyGenerator
- *
+ * @covers \SIL\CacheKeyProvider
  * @group semantic-interlanguage-links
  *
  * @license GNU GPL v2+
@@ -14,55 +13,49 @@ use SIL\CacheKeyGenerator;
  *
  * @author mwjames
  */
-class CacheKeyGeneratorTest extends \PHPUnit_Framework_TestCase {
+class CacheKeyProviderTest extends \PHPUnit_Framework_TestCase {
 
 	public function testCanConstruct() {
 
 		$this->assertInstanceOf(
-			'\SIL\CacheKeyGenerator',
-			new CacheKeyGenerator()
+			'\SIL\CacheKeyProvider',
+			new CacheKeyProvider()
 		);
 	}
 
 	public function testGetSiteCacheKey() {
 
-		$instance = new CacheKeyGenerator();
+		$instance = new CacheKeyProvider();
 
 		$this->assertSame(
-			$instance
-				->getSiteCacheKey( 'foo' ),
-			$instance
-				->getSiteCacheKey( 'foo' )
+			$instance->getSiteCacheKey( 'foo' ),
+			$instance->getSiteCacheKey( 'foo' )
 		);
 	}
 
 	public function testGetPageLanguageCacheBlobKey() {
 
-		$instance = new CacheKeyGenerator();
+		$instance = new CacheKeyProvider();
 
 		$this->assertSame(
-			$instance
-				->getPageLanguageCacheBlobKey( 'foo' ),
-			$instance
-				->getPageLanguageCacheBlobKey( 'foo' )
+			$instance->getPageLanguageCacheBlobKey( 'foo' ),
+			$instance->getPageLanguageCacheBlobKey( 'foo' )
 		);
 	}
 
 	public function testGetPageCacheKey() {
 
-		$instance = new CacheKeyGenerator();
+		$instance = new CacheKeyProvider();
 
 		$this->assertSame(
-			$instance
-				->getPageCacheKey( 'foo', true ),
-			$instance
-				->getPageCacheKey( 'foo', true )
+			$instance->getPageCacheKey( 'foo', true ),
+			$instance->getPageCacheKey( 'foo', true )
 		);
 	}
 
 	public function testPrefixModification() {
 
-		$instance = new CacheKeyGenerator();
+		$instance = new CacheKeyProvider();
 
 		$this->assertNotSame(
 			$instance